DonCheadle Re-vamped
Sorry for the prolonged silence. Work and illness have kept me super busy so, here is a more aesthetically please Don Cheadle. I removed Netstat until I figure out how to manipulate variables to look for Numbers greater than X. You'll notice I added service checks and version checks into the script. I'm currently working on the mess seen on lines 66 - 74. I would like to make it so that if the Variable is equal to X it gives me X output. If the Variable is equal to Y it gives me that output. I plan to do the same thing with the Dovecot or Courier checks once I figure out the issue.